Updating assignments/java/xor/Xor.java

+++ assignments/java/xor/Xor.java
@@ -25,7 +25,7 @@ public class Xor {
     String deciph = "";
     for (int i = 0; i<cipherBytes.length; i++) {
       cipherBytes[i] = (byte) Integer.parseInt(s.substring(2*i, 2*(i+1)),16);
-      System.out.println(cipherBytes[i] + " " + biteMe[i]);
+      System.out.println((int) cipherBytes[i] + " " + (int) biteMe[i]);
     }
 
     return "f";
